update coin func

+++ database/migrations/2018_02_18_071122_create_coins_table.php
@@ -19,16 +19,16 @@ class CreateCoinsTable extends Migration
             $table->string('coin_name');
             $table->string('coin_symbol');
             $table->integer('coin_rank');
-            $table->integer('price_usd');
-            $table->integer('price_btc');
+            $table->float('price_usd');
+            $table->float('price_btc');
             $table->bigInteger('volume_usd_24h');
             $table->bigInteger('market_cap_usd');
             $table->bigInteger('available_supply');
             $table->bigInteger('total_supply');
             $table->bigInteger('max_supply')->nullable();;
-            $table->integer('percent_change_1h');
-            $table->integer('percent_change_24h');
-            $table->integer('percent_change_7d');
+            $table->float('percent_change_1h');
+            $table->float('percent_change_24h');
+            $table->float('percent_change_7d');
             $table->timestamps();
         });
     }
